requirement.txt
flask==1.1.1
flask_script==2.0.6
flask_sqlalchemy==2.4.1
flask_migrate==2.5.2
python-dotenv==0.7.1
flask-json==0.3.4
gunicorn==19.7.1
supervisor==4.1.0